requirements.txt
pandas==2.0.0
pycryptodome==3.18.0
pymysql==1.0.3
python-dotenv==1.0.0
tqdm==4.65.0
